requirements.txt
# ===== Core ML =====
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
numpy>=1.24.0

# ===== API =====
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ===== Utilities =====
tqdm>=4.66.0
